+++ src/index.ts
@@ -140,11 +140,30 @@ export const options = {
       "Format the most popular Apex standard types. For example: `System`, `Map`, `DateTime`, `SObject` etc.",
   },
   apexFormatInlineComments: {
-    type: "boolean",
+    type: "choice",
     category: CATEGORY_APEX,
-    default: false,
-    description:
-      "Formats the inline comments to have a space and start from a capital letter.",
+    default: "none",
+    choices: [
+      {
+        value: "none",
+        description: "Do not apply any formatting to the inline comments.",
+      },
+      {
+        value: "spaced",
+        description: "Add a whitespace after the '//' in inline comment.",
+      },
+      {
+        value: "trimed",
+        description:
+          "Same as `spaced` but also remove all trailing whitespaces before and after the comment. Be aware that this will not work great with the commented-out code snippets.",
+      },
+      {
+        value: "strict",
+        description:
+          "Same as `trimed` but also make the first comment letter uppercase. Be aware that this will not work great with the commented-out code snippets.",
+      },
+    ],
+    description: "Formats the inline comments to be more readable.",
   },
   apexAnnotationsArgsSpacing: {
     type: "boolean",